Move quantized fixture to conftest.py
